feat(job): implement apply job

+++ app/src/main/java/com/oliver/siloker/data/network/model/response/JobAdResponseDto.kt
@@ -38,11 +38,11 @@ data class JobAdResponseItemDto(
 	val createdAt: String? = null,
 
 	@field:SerializedName("id")
-	val id: Int? = null,
+	val id: Long? = null,
 
 	@field:SerializedName("title")
 	val title: String? = null,
 
 	@field:SerializedName("employer_id")
-	val employerId: Int? = null
+	val employerId: Long? = null
 )
